+++ arctic_office_projects_api/main/faker/providers/grant/__init__.py
@@ -1,10 +1,14 @@
+import math
+
 from datetime import date
 from enum import Enum
+from typing import Optional
 
 from faker import Faker
 from faker.providers import BaseProvider
 # noinspection PyPackageRequirements
 from psycopg2.extras import DateRange
+from urllib.parse import quote_plus
 
 localized = False
 
@@ -55,8 +59,12 @@ class GrantCurrency(Enum):
         'iso_4217_code': 'EUR',
         'major_symbol': '€'
     }
-    USD = {
-        'iso_4217_code': 'USD',
+    NOK = {
+        'iso_4217_code': 'NOK',
+        'major_symbol': 'kr'
+    }
+    CAD = {
+        'iso_4217_code': 'CAD',
         'major_symbol': '$'
     }
 
@@ -99,9 +107,10 @@ class Provider(BaseProvider):
         Currently assumes:
           * 100% of UKRI standard/large grants are awarded in Pounds Sterling
           * 100% of EU standard grants are awarded in Euros
-          * 80% of Other grants are awarded in Pounds Sterling
+          * 87.5% of Other grants are awarded in Pounds Sterling
           * 10% of Other grants are awarded in Euros
-          * 10% of Other grants are awarded in US Dollars
+          * 2% of Other grants are awarded in Norwegian Krone
+          * 0.5% of Other grants are awarded in Canadian Dollars
 
         :type grant_type: GrantType
         :param grant_type: member of the GrantType enumerated class
@@ -110,37 +119,44 @@ class Provider(BaseProvider):
         :rtype: GrantCurrency
         :return: member of the GrantCurrency enumerated class, representing the currency of the funds for a grant
         """
-        if grant_type == GrantType.UKRI_LARGE_GRANT or grant_type == GrantType.UKRI_STANDARD_GRANT:
+        if grant_type is GrantType.UKRI_LARGE_GRANT or grant_type is GrantType.UKRI_STANDARD_GRANT:
             return GrantCurrency.GBP
-        if grant_type == GrantType.EU_STANDARD_GRANT:
+        if grant_type is GrantType.EU_STANDARD_GRANT:
             return GrantCurrency.EUR
 
         return GrantCurrency(self.random_element({
-            GrantCurrency.GBP: 0.80,
+            GrantCurrency.GBP: 0.875,
             GrantCurrency.EUR: 0.10,
-            GrantCurrency.USD: 0.10
+            GrantCurrency.NOK: 0.02,
+            GrantCurrency.CAD: 0.005
         }))
 
-    def grant_reference(self, grant_type: GrantType) -> str:
+    def grant_reference(self, grant_type: GrantType, ukri_council: Optional[UKRICouncil]) -> str:
         """
-        Generates a fake grant reference, based on the type of grant
-
-        Currently assumes all grant references are a 20 character random string
+        Generates a fake grant reference, based on the type of grant and optionally UKRI council
 
         :type grant_type: GrantType
         :param grant_type: member of the GrantType enumerated class
+        :type grant_type: UKRICouncils
+        :param grant_type: member of the UKRI councils enumerated class, if a UKRI GrantType
 
-        :example: 'ZmCRCKzQymLfZdNyXoJp'
+        :example: 'NE/A00123/1' (UKRI), '000123' (EU), 'ZmCRCKzQymLfZdNyXoJp' (Other)
         :rtype: str
         :return: fake grant reference
         """
-        lengths = {
-            'UKRI_STANDARD_GRANT': 20,
-            'UKRI_LARGE_GRANT': 20,
-            'EU_STANDARD_GRANT': 20,
-            'OTHER': 20
-        }
-        return self.faker.pystr(max_chars=lengths[grant_type.name])
+        if (grant_type is GrantType.UKRI_LARGE_GRANT and ukri_council is not None) or \
+                (grant_type is GrantType.UKRI_STANDARD_GRANT and ukri_council is not None):
+            version = self.random_element({
+                1: 0.98,
+                2: 0.016,
+                3: 0.004
+            })
+            return f"{ UKRICouncil(ukri_council).value['prefix'] }/{ self.faker.random_uppercase_letter() }" \
+                f"{ str(self.generator.random_int(min=1, max=999999)).zfill(5) }/{ version }"
+        elif grant_type is GrantType.EU_STANDARD_GRANT:
+            return str(self.generator.random_int(min=1, max=9999999)).zfill(6)
+
+        return self.faker.pystr(max_chars=20)
 
     def total_funds(self, grant_type: GrantType) -> int:
         """
@@ -160,18 +176,20 @@ class Provider(BaseProvider):
         :return: total amount of money awarded through a grant
         """
         grants_ranges = {
-            'UKRI_STANDARD_GRANT': (1000, 50000),
-            'UKRI_LARGE_GRANT': (50000, 1000000),
-            'EU_STANDARD_GRANT': (25000, 2000000),
-            'OTHER': (100, 25000)
+            'UKRI_STANDARD_GRANT': (1000, 80000),
+            'UKRI_LARGE_GRANT': (3000000, 3700000),
+            'EU_STANDARD_GRANT': (2000000, 10000000),
+            'OTHER': (100, 300000)
         }
-        return self.generator.random_int(
+        total_funds = self.generator.random_int(
             min=grants_ranges[grant_type.name][0],
             max=grants_ranges[grant_type.name][1]
         )
+        # Round to nearest 100
+        return int(math.floor(total_funds / 100.0)) * 100
 
     @staticmethod
-    def status(grant_duration: DateRange) -> GrantStatus:
+    def grant_status(grant_duration: DateRange) -> GrantStatus:
         """
         Determines the status of a grant based on its duration
 
@@ -189,3 +207,68 @@ class Provider(BaseProvider):
             return GrantStatus.Active
 
         return GrantStatus.Closed
+
+    def grant_funder(self, grant_type: GrantType) -> Optional[str]:
+        """
+        Determines the funder of a grant of project based on it's grant type (e.g. UKRI standard grant, EU grant, etc.)
+
+        Currently assumes:
+         * UKRI large and standard grants are funded:
+            * 3.4% by AHRC
+            * 1.9% by BBSRC
+            * 1.7% by EPSRC
+            * 1.7% by ESRC
+            * 0% by MRC
+            * 91% by NERC
+            * 0.3% by STFC
+         * EU grants are funded 100% by the EU
+         * 'other' grants:
+            * 90% by an existing random funder
+            * 10% by a new random funder
+
+        :type grant_type: GrantType
+        :param grant_type: member of the GrantType enumerated class
+
+        :example: 'NERC'
+        :rtype: str
+        :return: An identifier. For UKRI large/standard grants, an identifier for a UKRI council, for EU grants, an
+        identifier for the EU, for 'other' grants, any non-None value indicates an existing funder should be used.
+        """
+        if grant_type is GrantType.UKRI_LARGE_GRANT or grant_type is GrantType.UKRI_STANDARD_GRANT:
+            return UKRICouncil(self.random_element({
+                UKRICouncil.AHRC: 3.4,
+                UKRICouncil.BBSRC: 1.9,
+                UKRICouncil.EPSRC: 1.7,
+                UKRICouncil.ESRC: 1.7,
+                UKRICouncil.MRC: 0,
+                UKRICouncil.NERC: 91,
+                UKRICouncil.STFC: 0.3
+            })).name
+        elif grant_type is GrantType.EU_STANDARD_GRANT:
+            return 'EU'
+        elif grant_type is GrantType.OTHER:
+            return self.random_element({True: 0.8, None: 0.2})
+
+        return None
+
+    def grant_website(self, grant_type: GrantType, grant_reference: str) -> str:
+        """
+        Generates a fake grant website, based on the type of grant
+
+        :type grant_type: GrantType
+        :param grant_type: member of the GrantType enumerated class
+
+        :type grant_reference: str
+        :param grant_reference: fake grant reference
+
+        :example: 'https://gtr.ukri.org/projects?ref=ST%2FF006446%2F1' (UKRI),
+        'https://cordis.europa.eu/project/rcn/217700' (EU), 'https://www.example.com/123abc' (Other)
+        :rtype: str
+        :return: fake grant website
+        """
+        if grant_type is GrantType.UKRI_LARGE_GRANT or grant_type is GrantType.UKRI_STANDARD_GRANT:
+            return f"https://gtr.ukri.org/projects?ref={ quote_plus(grant_reference) }"
+        elif grant_type is GrantType.EU_STANDARD_GRANT:
+            return f"https://cordis.europa.eu/project/rcn/{ quote_plus(grant_reference) }"
+
+        return self.faker.uri()